Change approach to extract plugin context
